feat: allow disabling hyphenation replacement in escape helper
Previously, `escape(str, typographic: true)` did replace the hyphen
between word-characters with `"=` (i.e. the babel shorthand for a
hard-hyphen in german locale).

Now it is possible to disable the hyphenation replacement explicitly:

	<%= escape str, hyphenation: false %>

The default value for the `hyphenation` kwarg defaults to the value
given to the `typographic` kwarg (which in turn is true by default).